requirements_dev.txt
dicttoxml==1.6.6
Django==1.9.4
django-annoying==0.9.0
django-appconf==1.0.1
six==1.10.0
coverage==4.0.3
djangorestframework==3.3.3
djangorestframework-xml==1.3.0
Pillow==2.7.0
django-crispy-forms==1.6.0
django-filter==0.13.0
-e git://github.com/lotube/lotube-crawler.git#egg=lotube-crawler
